--- src/features/ui/media-slider/index.ts
@@ -1,1 +1,0 @@
-export * from "./slot-slider-with-upload";
