+++ src/features/ui/sliders/media-carousel/index.ts
@@ -1,0 +1,2 @@
+export * from "./media-carousel";
+export * from "./types";
