+++ src/features/ui/sliders/media-carousel/views/unknown.tsx
@@ -9,7 +9,7 @@ export function UnknownSlide({
   title?: string;
 }) {
   return (
-    <div className="relative flex h-full items-center justify-center rounded-lg bg-secondary text-primary">
+    <div className="bg-secondary text-primary relative flex h-full items-center justify-center rounded-lg">
       <FileQuestionIcon className="h-10 w-10" />
       <Caption text={title} />
     </div>
@@ -26,7 +26,7 @@ export function UnknownThumbnail({
   return (
     <div
       className={cn(
-        "flex h-full flex-col items-center justify-center gap-2 rounded-lg border-2 border-transparent bg-secondary px-2 text-xs text-primary",
+        "bg-secondary text-primary flex h-full flex-col items-center justify-center gap-2 rounded-lg border-2 border-transparent px-2 text-xs",
         isActive && "border-brand-500"
       )}
     >
