+++ src/shared/ui/sliders/index.ts
@@ -1,2 +1,1 @@
 export * from "./slide-view";
-export * from "./slot-slider";
