--- src/shared/ui/sliders/slot-slider/index.ts
@@ -1,2 +1,0 @@
-export * from "./slot-slider";
-export { type MediaType } from "./types";
